벽함정, 장판구역 추가

+++ Assets/Script/Eunsu/GameScene/GameManager.cs
@@ -107,7 +107,7 @@ public class GameManager : MonoBehaviour
 
     public void PlayerHPUIUpdate(int curHp, int maxHp)
     {
-        playerHPSlider.value = curHp / maxHp;
+        playerHPSlider.value = (float) curHp / maxHp;
         txtPlayerHP.text = $"{curHp} / {maxHp}";
     }  
 }
